requirements.txt
# Core
flask>=3.0
flask-sqlalchemy>=3.1
sqlalchemy>=2.0

# CLI
click>=8.1
rich>=13.0

# Config
pyyaml>=6.0
python-dotenv>=1.0

# Git integration
gitpython>=3.1

# Dev / Quality
pytest>=8.0
black>=24.0
flake8>=7.0
